Add TypeScript CodeQL Analysis (#4)
# Pull Request

## Description

This pull request includes updates to the
`.github/workflows/code-checks.yml` file to enhance the CodeQL analysis
workflow. The most important changes are the addition of a strategy
matrix to support multiple languages and the update to the repository
checkout action.

Enhancements to CodeQL analysis workflow:

* Added `statuses: write` permission to the `permissions` section.
* Introduced a strategy matrix to support analysis for both `typescript`
and `actions` languages.
* Updated the `Checkout repository` step to use
`actions/checkout@v4.2.2` with proper capitalization.
* Modified the `Initialize CodeQL` step to use the matrix language
variable for the `languages` input.