Add Python Dependabot Config (#9)
# Pull Request

## Description

This pull request includes an update to the `.github/dependabot.yml`
file to add a new configuration for the `pip` package ecosystem. The
most important change is the addition of a new update configuration for
`pip` dependencies in the `tests` directory.

Dependency management:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R34-R48):
Added a new `pip` package ecosystem configuration for the `tests`
directory, scheduled to run daily at 01:00 in the Europe/London
timezone, targeting the `main` branch, and including both patch and
minor updates.