Change Background Colours (#31)
# Pull Request

## Description

This pull request includes several updates to the styling of the `App`
and `LinkList` components in the `website` directory. The most important
changes include updating the background and text colors, modifying the
padding and layout, and removing unused CSS classes.

Styling updates:

*
[`website/src/components/App.tsx`](diffhunk://#diff-239a3a724c5261ff8313b35301e2eb4194e0ade5267ae23fbd74ed452b3959a1L6-R15):
Changed the background color of the main container to a gradient red and
updated the header's background color to `bg-stone-300`. The padding of
the header was also adjusted to use inline styles. The title text was
updated from "Jack Plowman" to "Project Links".
*
[`website/src/components/LinkList.tsx`](diffhunk://#diff-8d619ea7bd4bac792291acec9dda31b3dc721f8f5b82416414e98ef1d8aa4e96L40-R48):
Updated the background color of the link items to `bg-stone-100` and the
text color to `text-stone-800`.

Removal of unused CSS:

*
[`website/src/index.css`](diffhunk://#diff-6e9722f89187b8925b74f709b685ec38d1ddaf2e51bd58423903874dd16c0f6cL3-L5):
Removed the `.custom-padding` class as it is no longer used in the
updated components.