Bump GitHub Action Versions (#47)
# Pull Request

## Description

This pull request updates the `setup-uv` GitHub Action to version 6.0.1
across multiple workflow files to ensure compatibility with the latest
features and improvements.

### Dependency updates:

*
[`.github/actions/setup-test-dependencies/action.yml`](diffhunk://#diff-ee1e9a84357cfd23d067400706d92b6bb9d8fd8744e81563d9858e8c344df144L10-R10):
Updated the `setup-uv` action from version 5.4.2 to 6.0.1 for Python and
UV installation.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L171-R171):
Updated the `setup-uv` action from version 5.4.2 to 6.0.1 in two
separate jobs for installing the latest version of UV.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L171-R171)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L194-R194)