Add CodeLimit Job (#48)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
integrate the CodeLimit tool for repository analysis. The most important
change is the addition of the `run-codelimit` job in the
`.github/workflows/code-checks.yml` file.

### GitHub Actions Workflow Update:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R222-R236):
Added a new `run-codelimit` job to execute the CodeLimit analysis tool.
This includes setting up permissions, checking out the repository, and
running the `getcodelimit/codelimit-action` action.