Deploy to GitHub Pages (#10)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to automate
deployment to GitHub Pages. The workflow is triggered by pushes to the
`main` branch, manual dispatch, or a scheduled cron job.

### New GitHub Actions Workflow for Deployment:

* `.github/workflows/deploy.yml`: Added a workflow named "Deploy to
GitHub Pages" with two jobs:
-
[`build-site`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R1-R47):
Checks out the repository, builds the site using the `withastro/action`,
and prepares it for deployment.
-
[`deploy`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R1-R47):
Deploys the built site to GitHub Pages using the `actions/deploy-pages`
action. This job is dependent on the `build-site` job.