Add Link Checking Tests (#12)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
perform link testing after deployment. The new job ensures that all
links on the deployed site are functional.

Workflow enhancements:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R48-R65):
Introduced a `link-tests` job that runs after the deployment to GitHub
Pages. It uses the `linkinator-action` to verify the validity of links
on the deployed site (`https://jackplowman.github.io/project-links`)
with recursive checking and a timeout of 1000ms.