Fix Deploy Workflow 2 (#56)
# Pull Request

## Description

This pull request introduces a streamlined deployment workflow and a
minor configuration update for the website. The most significant changes
include merging the build and deploy jobs into a single job in the
GitHub Actions workflow and updating the Vite configuration to set a
base path for the project.

### Workflow improvements:
*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L17-R49):
Combined the `build-site` and `deploy` jobs into a single
`build-and-deploy` job to simplify the workflow. Added steps for setting
up dependencies, building the website, configuring GitHub Pages, and
uploading artifacts. Updated the `link-tests` job to depend on
`build-and-deploy` instead of `deploy`.

### Configuration updates:
*
[`website/vite.config.ts`](diffhunk://#diff-230359fd30686ac9bcb5f851d99fd24dd423628039d8b82dd52b99f09fb02270R8):
Added a `base` property to the Vite configuration to set the base path
for the project as `/project-links/`.