Run UI Tests on Deployment (#11)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow to
automate UI testing using Playwright. The most important change is the
addition of the `ui-tests` job, which ensures UI tests are run after the
build-and-deploy process.

### Workflow enhancements:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3R46-R63):
Added a new `ui-tests` job that runs on `ubuntu-latest`. This job
includes steps for checking out the repository, setting up Python
dependencies, installing Playwright dependencies, and running UI tests
using predefined `just` commands. The job is dependent on the successful
completion of the `build-and-deploy` job.