Add website sitemap (#66)
# Pull Request

## Description

This pull request introduces the `vite-plugin-sitemap` package to the
project, enabling automatic sitemap generation for the website. The
changes include updates to dependencies, configuration adjustments, and
integration of the plugin into the Vite build process.

### Dependency Updates:
* Added `vite-plugin-sitemap` version `0.7.1` to both `dependencies` in
`website/package.json` and `node_modules` in
`website/package-lock.json`.
[[1]](diffhunk://#diff-fae242fbf77a8a9d52625664bc36ea12316586f8a716b41137d897a2b7e3df76L35-R36)
[[2]](diffhunk://#diff-999051457a1934c824a5f8395589ea5e749a581e9f70a43585c885319cb56e64L33-R34)
[[3]](diffhunk://#diff-999051457a1934c824a5f8395589ea5e749a581e9f70a43585c885319cb56e64R6312-R6317)

### Configuration Changes:
* Updated `website/vite.config.ts` to include `vite-plugin-sitemap` in
the Vite plugins array, ensuring the sitemap is generated during the
build process.