Add Sitemap Test (#67)
# Pull Request

## Description

This pull request adds a new test to verify the sitemap functionality of
the website. It introduces a new dependency and includes assertions to
ensure the sitemap is accessible and correctly formatted.

### Additions to testing:
*
[`tests/ui/test_metadata.py`](diffhunk://#diff-1e7e86c0ec0c55344ec4f7908c622d1271b7b5a0e853cb5bb347a11509dc2fe0R15-R26):
Added a new `test_sitemap` function to validate the sitemap's
availability, status code, content type, and structure.

### Dependency updates:
*
[`tests/ui/test_metadata.py`](diffhunk://#diff-1e7e86c0ec0c55344ec4f7908c622d1271b7b5a0e853cb5bb347a11509dc2fe0R2):
Imported the `get` function from the `requests` library to facilitate
HTTP requests for the sitemap test.