Remove Pinact Configuration (#86)
# Pull Request

## Description

This pull request removes the configuration file for the `pinact` tool
and updates related commands in the `Justfile` to no longer reference
the deleted configuration file.

### Removal of `pinact` configuration:
* Deleted the `.github/other-configurations/pinact.yml` file, which
contained the `pinact` configuration, including the version and ignored
actions.

### Updates to `Justfile`:
* Updated the `pinact` commands (`pinact-run`, `pinact-check`, and
`pinact-update`) in the `Justfile` to remove references to the deleted
`pinact.yml` configuration file. These commands now run without
specifying a custom configuration file.