Swap linkinator to linkcheck (#152)
# Pull Request

## Description

This pull request updates the link-checking tool used in the deployment
workflow to a newer version with a different action. The most important
change is the replacement of the `linkinator-action` with `linkcheck`,
along with adjustments to its configuration.

Workflow updates:

*
[`.github/workflows/deploy.yml`](diffhunk://#diff-28802fbf11c83a2eee09623fb192785e7ca92a3f40602a517c011b947a1822d3L83-R85):
Replaced `JustinBeckwith/linkinator-action` (v1.11.0) with
`filiph/linkcheck` (v3.0.0) for link testing. Updated the configuration
to use the `arguments` parameter instead of `paths`, and removed unused
parameters (`recurse`, `timeout`, `markdown`).